+++ lib/rubocop/cop/style/documentation.rb
@@ -30,6 +30,7 @@ module RuboCop
         MSG = 'Missing top-level %<type>s documentation comment.'.freeze
 
         def_node_matcher :constant_definition?, '{class module casgn}'
+        def_node_search :outer_module, '(const (const nil? _) _)'
 
         def on_class(node)
           _, _, body = *node
@@ -50,6 +51,8 @@ module RuboCop
         def check(node, body, type)
           return if namespace?(body)
           return if documentation_comment?(node) || nodoc_comment?(node)
+          return if compact_namespace?(node) &&
+                    nodoc_comment?(outer_module(node).first)
 
           add_offense(node,
                       location: :keyword,
@@ -64,6 +67,10 @@ module RuboCop
           else
             constant_definition?(node)
           end
+        end
+
+        def compact_namespace?(node)
+          node.loc.name.source =~ /::/
         end
 
         # First checks if the :nodoc: comment is associated with the